+++ lib/cocoapods/command/spec.rb
@@ -69,7 +69,8 @@ module Pod
            ['--no-subspecs', 'Lint skips validation of subspecs'],
            ['--no-clean',    'Lint leaves the build directory intact for inspection'],
            ['--sources=https://github.com/artsy/Specs', 'The sources to pull dependant pods from ' \
-            '(defaults to https://github.com/CocoaPods/Specs.git)']].concat(super)
+            '(defaults to https://github.com/CocoaPods/Specs.git)'],
+           ['--private', 'Lint skips checks that apply only to public specs']].concat(super)
         end
 
         def initialize(argv)
@@ -79,6 +80,7 @@ module Pod
           @subspecs     = argv.flag?('subspecs', true)
           @only_subspec = argv.option('subspec')
           @source_urls  = argv.option('sources', 'https://github.com/CocoaPods/Specs.git').split(',')
+          @private      = argv.flag?('private', false)
           @podspecs_paths = argv.arguments!
           super
         end
@@ -87,7 +89,7 @@ module Pod
           UI.puts
           invalid_count = 0
           podspecs_to_lint.each do |podspec|
-            validator             = Validator.new(podspec, @source_urls)
+            validator             = Validator.new(podspec, @source_urls, @private)
             validator.quick       = @quick
             validator.no_clean    = !@clean
             validator.only_errors = @only_errors